+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Tags can now be hidden from the navigation bar in the HTML report by setting the `showInNavigation` attribute to `false` [#211](https://github.com/TNG/JGiven/issues/211).
 * Added a new CurrentScenario interface similar to CurrentStep.
 * The CurrentScenario interface allows adding tags programmatically, see [#172](https://github.com/TNG/JGiven/issues/172).
+* Allow tag annotations on step methods and step classes.
 
 ## Breaking Changes in the JSON model